update(Twitter): Made some features to use flags
