+++ src/main/kotlin/crimera/patches/twitter/featureFlag/RemoveViewCountPatch.kt
@@ -1,37 +1,30 @@
-package crimera.patches.twitter.misc.viewcount
+package crimera.patches.twitter.featureFlag
 
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
-import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.patch.BytecodePatch
-import app.revanced.patcher.patch.PatchException
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
+import crimera.patches.twitter.featureFlag.fingerprints.FeatureFlagLoadFingerprint
 import crimera.patches.twitter.misc.settings.SettingsPatch
 import crimera.patches.twitter.misc.settings.fingerprints.SettingsStatusLoadFingerprint
-import crimera.patches.twitter.misc.viewcount.fingerprints.RemoveViewCountPatchFingerprint
 
 // Credits to @iKirby
 @Patch(
     name = "Remove view count",
     description = "Removes the view count from the bottom of tweets",
+    dependencies = [SettingsPatch::class,FeatureFlagPatch::class],
     compatiblePackages = [CompatiblePackage("com.twitter.android")],
 )
 @Suppress("unused")
 object RemoveViewCountPatch: BytecodePatch(
-    setOf(RemoveViewCountPatchFingerprint)
+    setOf(FeatureFlagLoadFingerprint,SettingsStatusLoadFingerprint)
 ) {
     override fun execute(context: BytecodeContext) {
-        val result = RemoveViewCountPatchFingerprint.result
-            ?: throw PatchException("Fingerprint not found")
-
-        val method = result.mutableMethod
-
-        method.addInstructions(0, """
-            invoke-static {}, ${SettingsPatch.PREF_DESCRIPTOR};->hideViewCount()Z
-            move-result v0
-            return v0
-        """.trimIndent())
+        FeatureFlagLoadFingerprint.result!!.mutableMethod.addInstruction(
+            0,
+            "${SettingsPatch.FSTS_DESCRIPTOR}->viewCount()V"
+        )
 
         SettingsStatusLoadFingerprint.result!!.mutableMethod.addInstruction(
             0,
